:recycle: Move transaction validation to different file and add type 4 check
